Testing: structured arguments
This should make it a bit easier to manage argument overrides than the
set of templates we have now, especially while we wait for
componentconfig to be everywhere.

Arguments can be defaulted, and those defaults can be overriden or
appended to, and then finally render to a slice of strings to be passed
as arguments and such.

This should make it easier to configure the API server without needing
to splat out the existing default arguments, maybe splice some stuff
out, etc.

For example:

```go
apiServer.Configure().
  Disable("insecure-port").
  Append("vmodule", "httplog.go=10").
  Set("v", "5")
```